refactor : remove unused imported module in category service logic
