+++ src/routes/index.ts
@@ -2,7 +2,7 @@ import express from 'express';
 
 import userRoute from './user/user.route';
 import mediaRoute from './media.route';
-import categoryRoute from './category.route';
+import categoryRoute from './category/category.route';
 import sessionRouter from './session/session.route';
 
 const router = express.Router();
