+++ src/validations/category.validation.ts
@@ -1,5 +1,3 @@
-import joi from 'joi';
 import { RequestSchema } from '../middleware/validate';
-import { pageNationQuerySchema } from './pagenation.validation';
 
 export const getCategories: RequestSchema = {};
